cleanup the code
